snyk vulnerability update

+++ app/meal-list.component.ts
@@ -17,6 +17,9 @@ export class MealListComponent implements OnInit {
   selectedMeal: Meal;
   restaurants: Restaurant[];
   selectedRestaurant = 0;
+  badData = false;
+  badDataMessage1 = '';
+  badDataMessage2 = '';
 
   constructor(private router: Router,
               private mealService: MealService,
@@ -46,10 +49,42 @@ export class MealListComponent implements OnInit {
     }
   }
 
-  save(amount: string, newtype: string, newdate: string, oldtype: string, olddate: string) {
+  save(amount: string, newtype: string, newdate: string,
+       oldtype: string, olddate: string): void {
     let mydate = moment(newdate, 'MM/DD/YYYY');
+    let amtAsNum = parseFloat(amount);
+
+    this.badDataMessage1 = '';
+    this.badDataMessage2 = '';
+    this.badData = false;
+
+    if (isNaN(amtAsNum)) {
+      this.badDataMessage1 = "Please enter a numeric amount.";
+      this.badData = true;
+    }
+    else {
+      if (amtAsNum <= 0) {
+        this.badDataMessage1 = "Please enter an amount greater than zero.";
+        this.badData = true;
+      }
+    }
+    if (!newdate) {
+      this.badDataMessage2 = "Please enter a date.";
+      this.badData = true;
+    }
+    else {
+      if (!mydate.isValid()) {
+        this.badDataMessage2 = "Please enter a valid date.";
+        this.badData = true;
+      }
+    }
+
+    if (!this.badData) {
+      return null;
+    }
 
     newdate = mydate.format('YYYYMMDD');
+
     console.log('delete before save:', olddate, oldtype, this.selectedRestaurant);
     this.mealService.delete(olddate, oldtype, this.selectedRestaurant);
